6 Direction Combination & Experimental Deletion
Created new static class for the creation and 6 direction combination of modular blocks.  A modular block will now have  a reference to each other adjacent modular block and vice versa. 

State of deletion:
Blocks and the reference of each connected block will be removed. 
Problems: If a block is deleted and it had other modular blocks as children, those will also be deleted (reference is wonky then).